+++ Vault/Sources/VaultiOS/Views/Detail/OTP/OTPCodeDetailView.swift
@@ -80,6 +80,7 @@ struct OTPCodeDetailView<PreviewGenerator: VaultItemPreviewViewGenerator<VaultIt
             isShowingDeleteConfirmation: $isShowingDeleteConfirmation,
             navigationPath: $navigationPath,
             presentationMode: presentationMode,
+            persistedLockState: viewModel.editingModel.initialDetail.lockState,
         ) {
             if viewModel.isInEditMode {
                 if viewModel.showsKeyEditingFields {